+++ packages/gestalt/src/TextField/TextField.js
@@ -9,6 +9,7 @@ import styles from './TextField.css';
 type State = {
   focused: boolean,
   errorIsOpen: boolean,
+  errorMessage?: string,
 };
 
 type Props = {|
@@ -63,16 +64,21 @@ export default class TextField extends React.Component<Props, State> {
     type: 'text',
   };
 
-  state: State = {
+  static getDerivedStateFromProps(props: Props, state: State) {
+    if (props.errorMessage !== state.errorMessage) {
+      return {
+        errorIsOpen: !!props.errorMessage,
+        errorMessage: props.errorMessage,
+      };
+    }
+
+    return null;
+  }
+
+  state = {
     focused: false,
     errorIsOpen: false,
   };
-
-  componentWillReceiveProps(nextProps: Props) {
-    if (nextProps.errorMessage !== this.props.errorMessage) {
-      this.setState({ errorIsOpen: !!nextProps.errorMessage });
-    }
-  }
 
   textfield: ?HTMLElement;